:art: Reorganization of the project structure
