+++ src/main/java/blue/lhf/bytepaper/library/syntax/event/player/EventPlayerCommandPreprocess.java
@@ -1,4 +1,4 @@
-package blue.lhf.bytepaper.library.syntax.player;
+package blue.lhf.bytepaper.library.syntax.event.player;
 
 import org.bukkit.entity.Player;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
